Hide Recently Added Items (+ more cleanup).
Only display Recently added items if they are set to display in
the theme options. Currently there is no way to set this in the
admin interface, but we don't want these displayed by default
so this will work for now.

Cleaned up some markup and removed a div close that was unmatched
when most of the default omeka frontpage displays are hidden.