Split elements into separate divs for display
Eventually each of these divs will be made into a tab.
Currently if an element is not in our list here it is not displayed.
We will probably want to keep track of what has not been displayed
yet and dump it into the 3rd div, the more-info tab.

I'm pretty certain we need to add a few fields that apply to People
and Moving Image item types, but this should cover Still Images
based on the mockup I was given.