Fix person about info w/ multiple cultures
Was displaying them side by side when they should be stacked
vertically. Basically this resulted from moving the about
info into #item-top on person pages when divs in it all get
vertical-align top. The simplest solution is to only apply that
vertical alignment to the immediate child divs of #item-top.